+++ examples/volumetric/numpy2volume2.py
@@ -7,7 +7,7 @@ data_matrix[0:30,   0:30,  0:30] = 1
 data_matrix[30:50, 30:60, 30:70] = 2
 data_matrix[50:69, 60:79, 70:89] = 3
 
-vol = Volume(data_matrix, c=['white','b','g','r'], mapper='gpu')
+vol = Volume(data_matrix, c=['white','b','g','r'], mode=1)
 vol.add_scalarbar3d()
 vol.print()